requirements.txt
opencv-python
Pillow
scikit-learn
labelme
numpy
